requirements.txt
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==6.2.0
python-decouple==3.6
django-rest-knox==4.2.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
Pillow==9.1.1